Feat/animation (#10)
* Updated probe to include per-output information about PTS

* Added interpolation support for most of the types

* animations are beta

* Added api for animations

* Added an anim remover

* Added a basic script and fixed decoding of node pointers

+++ src/vortex/ui/value.h
@@ -94,7 +94,7 @@ struct value_traits<uintptr_t> {
         if (list.GetType(index) != VTYPE_DOUBLE) {
             return false;
         }
-        out = std::bit_cast<double>(list.GetDouble(index));
+        out = std::bit_cast<uintptr_t>(list.GetDouble(index));
         return true;
     }
 };
